geotiff: extract CRS resolution helpers to _crs.py (#1881)
Step 3 of #1813's multi-PR refactor of __init__.py. Pure code motion;
no public API change.

Moved into a new xrspatial/geotiff/_crs.py:

- _wkt_to_epsg: best-effort EPSG extraction from WKT/PROJ, with the
  strict-mode-aware fallback warning.
- _resolve_crs_to_wkt: normalises int EPSG / WKT / PROJ to a canonical
  WKT string for downstream writers and VRT serialisation. Used by
  to_geotiff, write_geotiff_gpu, and write_vrt.

Both helpers depend only on pyproj (lazy-imported inside) and on
GeoTIFFFallbackWarning / _geotiff_strict_mode, which moved to
_runtime in #1890. __init__.py re-imports them so every existing
caller keeps working.

Net __init__.py change: 4836 -> 4726 lines (-110). _crs.py is 128
lines.

Verification: pixel-parity matrix from #1889, runtime-sentinel
identity tests from #1890, and the existing CRS test files
(test_write_vrt_crs_1715.py, test_user_defined_crs_wkt_1632.py,
test_wkt_only_crs_warning_1768.py) all pass (249/249).

Refs #1813.
